Removendo redudâncias das models , refatorando camadas , camada usuario iniciada, Camada de Exceptions criada, controller generico criado

+++ src/main/java/br/com/senac/ado/zoologico/service/UsuarioService.java
@@ -2,9 +2,12 @@ package br.com.senac.ado.zoologico.service;
 
 import br.com.senac.ado.zoologico.dto.UsuarioDTO;
 import br.com.senac.ado.zoologico.entity.Usuario;
+import br.com.senac.ado.zoologico.exception.ConflictException;
+import br.com.senac.ado.zoologico.exception.ResourceNotFoundException;
 import br.com.senac.ado.zoologico.repository.UsuarioRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.ResourceAccessException;
 
 import java.util.List;
 import java.util.UUID;
@@ -15,42 +18,55 @@ public class UsuarioService {
 
     private final UsuarioRepository repository;
 
-    public List<Usuario> listarTodos() {
+    public List<Usuario> getAll() {
         return repository.findAll();
     }
 
-    public Usuario buscar(UUID id) {
-        return repository.findById(id).orElse(null);
+    public Usuario findById(UUID id) {
+        return repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não Encontrado com o ID informado"));
     }
 
-    public Usuario registrar(UsuarioDTO dto) {
-        if (repository.findByUsername(dto.getUsername()).isPresent()) {
-            throw new RuntimeException("Usuário já existe");
-        }
+    public UUID save(UsuarioDTO dto) {
+
+       repository.findByEmail(dto.getEmail()).ifPresent(user -> {
+          throw new ConflictException("Usuario já existe: " + dto.getUsername());
+       });
 
         Usuario user = new Usuario();
         user.setUsername(dto.getUsername());
+        user.setEmail(dto.getEmail());
         user.setSenha(dto.getSenha());
         user.setPapel(dto.getPapel());
-        return repository.save(user);
+        return repository.save(user).getId();
     }
 
-    public Usuario atualizar(UUID id, Usuario dto) {
+    public void update(UUID id, Usuario dto) {
+
+
         Usuario user = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException("Recurso não Encontrado com o ID informado"));
+
+        if (repository.existsByEmailAndIdNot(dto.getEmail(), id)) {
+            throw new ConflictException("Conflito de dados , tente novamente");
+        }
+
         user.setUsername(dto.getUsername());
+        user.setEmail(dto.getEmail());
         user.setSenha(dto.getSenha());
         user.setPapel(dto.getPapel());
-        return repository.save(user);
+        repository.save(user);
     }
 
-    public void excluir(UUID id) {
+    public void deleteById(UUID id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado para exclusão");
+        }
         repository.deleteById(id);
     }
 
-    public boolean autenticar(String username, String senha) {
-        Usuario user = repository.findByUsername(username)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
-        return user.getSenha().equals(senha);
-    }
+//    public boolean autenticar(String username, String senha) {
+//        Usuario user = repository.findByEmail(username)
+//                .orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado"));
+//        return user.getSenha().equals(senha);
+//    }
 }
